# Optimize FileInfo(Version) transfer