--- a/cmd/storage-rest-client.go
+++ b/cmd/storage-rest-client.go
@@ -35,6 +35,7 @@ import (
 	"github.com/minio/minio/cmd/logger"
 	"github.com/minio/minio/cmd/rest"
 	xnet "github.com/minio/minio/pkg/net"
+	"github.com/tinylib/msgp/msgp"
 )
 
 func isNetworkError(err error) bool {
@@ -291,7 +292,7 @@ func (client *storageRESTClient) WriteMetadata(ctx context.Context, volume, path
 	values.Set(storageRESTFilePath, path)
 
 	var reader bytes.Buffer
-	if err := gob.NewEncoder(&reader).Encode(fi); err != nil {
+	if err := msgp.Encode(&reader, &fi); err != nil {
 		return err
 	}
 
@@ -306,7 +307,7 @@ func (client *storageRESTClient) DeleteVersion(ctx context.Context, volume, path
 	values.Set(storageRESTFilePath, path)
 
 	var buffer bytes.Buffer
-	if err := gob.NewEncoder(&buffer).Encode(fi); err != nil {
+	if err := msgp.Encode(&buffer, &fi); err != nil {
 		return err
 	}
 
@@ -342,7 +343,7 @@ func (client *storageRESTClient) CheckParts(ctx context.Context, volume string, 
 	values.Set(storageRESTFilePath, path)
 
 	var reader bytes.Buffer
-	if err := gob.NewEncoder(&reader).Encode(fi); err != nil {
+	if err := msgp.Encode(&reader, &fi); err != nil {
 		logger.LogIf(context.Background(), err)
 		return err
 	}
@@ -378,8 +379,7 @@ func (client *storageRESTClient) ReadVersion(ctx context.Context, volume, path, 
 		return fi, err
 	}
 	defer http.DrainBody(respBody)
-
-	err = gob.NewDecoder(respBody).Decode(&fi)
+	err = msgp.Decode(respBody, &fi)
 	return fi, err
 }
 
@@ -448,10 +448,10 @@ func (client *storageRESTClient) WalkSplunk(ctx context.Context, volume, dirPath
 		defer close(ch)
 		defer http.DrainBody(respBody)
 
-		decoder := gob.NewDecoder(respBody)
+		decoder := msgp.NewReader(respBody)
 		for {
 			var fi FileInfo
-			if gerr := decoder.Decode(&fi); gerr != nil {
+			if gerr := fi.DecodeMsg(decoder); gerr != nil {
 				// Upon error return
 				return
 			}
@@ -483,10 +483,10 @@ func (client *storageRESTClient) WalkVersions(ctx context.Context, volume, dirPa
 		defer close(ch)
 		defer http.DrainBody(respBody)
 
-		decoder := gob.NewDecoder(respBody)
+		decoder := msgp.NewReader(respBody)
 		for {
 			var fi FileInfoVersions
-			if gerr := decoder.Decode(&fi); gerr != nil {
+			if gerr := fi.DecodeMsg(decoder); gerr != nil {
 				// Upon error return
 				if gerr != io.EOF {
 					logger.LogIf(GlobalContext, gerr)
@@ -520,10 +520,10 @@ func (client *storageRESTClient) Walk(ctx context.Context, volume, dirPath, mark
 		defer close(ch)
 		defer http.DrainBody(respBody)
 
-		decoder := gob.NewDecoder(respBody)
+		decoder := msgp.NewReader(respBody)
 		for {
 			var fi FileInfo
-			if gerr := decoder.Decode(&fi); gerr != nil {
+			if gerr := fi.DecodeMsg(decoder); gerr != nil {
 				// Upon error return
 				return
 			}
@@ -575,9 +575,9 @@ func (client *storageRESTClient) DeleteVersions(ctx context.Context, volume stri
 	values.Set(storageRESTTotalVersions, strconv.Itoa(len(versions)))
 
 	var buffer bytes.Buffer
-	encoder := gob.NewEncoder(&buffer)
+	encoder := msgp.NewWriter(&buffer)
 	for _, version := range versions {
-		encoder.Encode(&version)
+		version.EncodeMsg(encoder)
 	}
 
 	errs = make([]error, len(versions))
@@ -632,7 +632,7 @@ func (client *storageRESTClient) VerifyFile(ctx context.Context, volume, path st
 	values.Set(storageRESTFilePath, path)
 
 	var reader bytes.Buffer
-	if err := gob.NewEncoder(&reader).Encode(fi); err != nil {
+	if err := msgp.Encode(&reader, &fi); err != nil {
 		return err
 	}