… cmd/storage-datatypes_gen.go
… cmd/storage-datatypes_gen_test.go