## Motivation and Context

File Info decoding in particular is showing up as a major allocator and time consumer.

We should bump internal storage REST version as well. Maybe add this as part of a release that updates it anyway.

Switch to messagepack for cross-server transfers:

```
MSGP:

Size: 945 bytes

BenchmarkEncodeFileInfoMsgp-32    	 1558444	       866 ns/op	   1.16 MB/s	       0 B/op	       0 allocs/op
BenchmarkDecodeFileInfoMsgp-32    	  479968	      2487 ns/op	   0.40 MB/s	     848 B/op	      18 allocs/op

GOB (existing):

Size: 1409 bytes

BenchmarkEncodeFileInfoGOB-32    	  333339	      3237 ns/op	   0.31 MB/s	     576 B/op	      19 allocs/op
BenchmarkDecodeFileInfoGOB-32    	   20869	     57837 ns/op	   0.02 MB/s	   16439 B/op	     428 allocs/op
```

Benchmark included. MB/s is FileInfo objects per second (1 byte per operation).

## How to test this PR?

Standard server operation should cover this.

## Types of changes
- [x] New feature (non-breaking change which adds functionality)